gfdgdfg

+++ app/src/main/java/com/rongcloud/model/CarShop.java
@@ -82,7 +82,7 @@ public class CarShop implements Serializable{
     private String city;
     private String createTime;
     private double distance;
-    private double grade;
+    private int grade;
     private String id;
     private String latitude;
     private String logo;
@@ -150,11 +150,11 @@ public class CarShop implements Serializable{
         this.distance = distance;
     }
 
-    public double getGrade() {
+    public int getGrade() {
         return grade;
     }
 
-    public void setGrade(double grade) {
+    public void setGrade(int grade) {
         this.grade = grade;
     }
